重构Redission组件
