+++ aquarius-cache-aop/src/main/java/com/nepxion/aquarius/cache/constant/CacheConstant.java
@@ -14,11 +14,6 @@ public class CacheConstant {
 
     public static final String CACHE_TYPE_REDIS = "redisCache";
 
-    public static final String CACHE_PLUGIN = "cache.plugin";
-
-    public static final String CACHE_PLUGIN_REDIS = "redisPlugin";
-    public static final String CACHE_PLUGIN_REDISSON = "redissonPlugin";
-
     public static final String CACHE_AOP_EXCEPTION_IGNORE = "cache.aop.exception.ignore";
 
     public static final String CACHE_EXPIRE = "cache.expire";
