--- aquarius-cache-redis/src/main/java/com/nepxion/aquarius/cache/redis/condition/RedissonCachePluginCondition.java
@@ -1,19 +1,0 @@
-package com.nepxion.aquarius.cache.redis.condition;
-
-/**
- * <p>Title: Nepxion Aquarius</p>
- * <p>Description: Nepxion Aquarius</p>
- * <p>Copyright: Copyright (c) 2017-2050</p>
- * <p>Company: Nepxion</p>
- * @author Haojun Ren
- * @version 1.0
- */
-
-import com.nepxion.aquarius.cache.constant.CacheConstant;
-import com.nepxion.aquarius.common.condition.AquariusCondition;
-
-public class RedissonCachePluginCondition extends AquariusCondition {
-    public RedissonCachePluginCondition() {
-        super(CacheConstant.CACHE_PLUGIN, CacheConstant.CACHE_PLUGIN_REDISSON);
-    }
-}
