+++ aquarius-cache-redis/src/main/java/com/nepxion/aquarius/cache/redis/configuration/RedissonCacheConfiguration.java
@@ -1,0 +1,49 @@
+package com.nepxion.aquarius.cache.redis.configuration;
+
+/**
+ * <p>Title: Nepxion Aquarius</p>
+ * <p>Description: Nepxion Aquarius</p>
+ * <p>Copyright: Copyright (c) 2017-2050</p>
+ * <p>Company: Nepxion</p>
+ * @author Haojun Ren
+ * @version 1.0
+ */
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Conditional;
+import org.springframework.context.annotation.Configuration;
+
+import com.nepxion.aquarius.cache.CacheDelegate;
+import com.nepxion.aquarius.cache.redis.condition.RedisCacheCondition;
+import com.nepxion.aquarius.cache.redis.impl.RedissonCacheDelegateImpl;
+import com.nepxion.aquarius.common.redisson.adapter.RedissonAdapter;
+import com.nepxion.aquarius.common.redisson.constant.RedissonConstant;
+import com.nepxion.aquarius.common.redisson.handler.RedissonHandler;
+import com.nepxion.aquarius.common.redisson.handler.RedissonHandlerImpl;
+
+@Configuration
+public class RedissonCacheConfiguration {
+    @Value("${" + RedissonConstant.PATH + ":" + RedissonConstant.DEFAULT_PATH + "}")
+    private String redissonPath;
+
+    @Autowired(required = false)
+    private RedissonAdapter redissonAdapter;
+
+    @Bean
+    @Conditional(RedisCacheCondition.class)
+    public CacheDelegate redisCacheDelegate() {
+        return new RedissonCacheDelegateImpl();
+    }
+
+    @Bean
+    @Conditional(RedisCacheCondition.class)
+    public RedissonHandler redissonHandler() {
+        if (redissonAdapter != null) {
+            return redissonAdapter.getRedissonHandler();
+        }
+
+        return new RedissonHandlerImpl(redissonPath);
+    }
+}
